+++ main.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/cloudogu/k8s-backup-operator/pkg/metrics"
-	"github.com/cloudogu/k8s-backup-operator/pkg/ownerreference"
 	"github.com/cloudogu/k8s-backup-operator/pkg/provider"
 	blueprintv3 "github.com/cloudogu/k8s-blueprint-lib/v3/client"
 	doguv2Client "github.com/cloudogu/k8s-dogu-lib/v2/client"
@@ -322,11 +321,6 @@ func configureReconcilers(ctx context.Context, k8sManager controllerManager, ope
 		return fmt.Errorf("failed to update additional images in existing resources: %w", err)
 	}
 
-	ownerRefRecreator, err := ownerreference.NewRecreator(k8sManager.GetConfig(), namespace)
-	if err != nil {
-		return fmt.Errorf("unable to create owner reference client: %w", err)
-	}
-
 	requeueHandler := requeue.NewRequeueHandler(ecosystemClientSet, recorder, operatorConfig.Namespace)
 	cleanupManager := cleanup.NewManager(doguClient.Dogus(namespace))
 	restoreManager := restore.NewRestoreManager(
@@ -335,13 +329,12 @@ func configureReconcilers(ctx context.Context, k8sManager controllerManager, ope
 		operatorConfig.Namespace,
 		recorder,
 		cleanupManager,
-		ownerRefRecreator,
 	)
 	if err = (restore.NewRestoreReconciler(ecosystemClientSet, recorder, operatorConfig.Namespace, restoreManager, requeueHandler)).SetupWithManager(k8sManager); err != nil {
 		return fmt.Errorf("unable to create restore controller: %w", err)
 	}
 
-	backupManager := backup.NewBackupManager(k8sClient, ecosystemClientSet, blueprintClient, operatorConfig.Namespace, recorder, globalConfig, ownerRefRecreator)
+	backupManager := backup.NewBackupManager(k8sClient, ecosystemClientSet, blueprintClient, operatorConfig.Namespace, recorder, globalConfig)
 	if err = (backup.NewBackupReconciler(ecosystemClientSet, recorder, operatorConfig.Namespace, backupManager, requeueHandler)).SetupWithManager(k8sManager); err != nil {
 		return fmt.Errorf("unable to create backup controller: %w", err)
 	}