+++ pkg/restore/interfaces.go
@@ -109,7 +109,3 @@ type ecosystemV1Alpha1Interface interface {
 type restoreProvider interface {
 	provider.Provider
 }
-
-type ownerReferenceRestore interface {
-	RestoreOwnerReferences(ctx context.Context) error
-}